requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-allauth==0.24.1
django-cors-headers==1.1.0
django-friendship==1.3.1
django-toolbelt==0.0.1
gunicorn==19.4.1
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
static3==0.6.1
wheel==0.24.0
whitenoise==2.0.6
